extract refreshSkillsIfChanged to src/utils/skills-refresh.ts
Per PR #555 review: keep the helper + its types out of runner.ts so the
file doesn't keep growing. New module follows the same pattern as
src/utils/mcp-server-fetcher.ts (worker-side, HTTP-only, no DB imports).
Test now imports the real export instead of carrying a duplicate copy.
